If there's only one guess left, the strategy will guess a word.
It doesn't make much sense to guess a correct letter if that's
going to be the last move. :-P

I also fixed the maximum number of wrong guess, it was 25 and now
it's the correct 11 (although Wikipedia says 9, the sample games
provided by Factual play to 11).

Altered the way the number of games are handled, it will now
replay some games if there are more games to be played than
solution words.